straighten() Method
Method for robot to turn towards a certain compass sensor value.  This requires the input to be in terms of the starting compass value (measured before any movement in the program about 14 milliseconds after the program starts running so the compass can get a reading) in order to go in that direction.  Say you want the robot to be 45 degrees to the right of the starting position, or vertical north, and the variable firstCompassSensor holds the first reading of the compass sensor.  When calling the method, you should put:

straighten(firstCompassSensor + 45);

In theory, it should work out ok ;)